+++ ichnaea/webapp/monitor.py
@@ -1,32 +1,15 @@
 """
 Heartbeat and monitor views to check service and backend health.
 """
-import os.path
-import socket
 import time
 
-import pkg_resources
 from pyramid.httpexceptions import HTTPServiceUnavailable
-import simplejson
 
-from ichnaea import ROOT
+from ichnaea.config import (
+    LOCAL_FQDN,
+    VERSION_INFO,
+)
 from ichnaea.webapp.view import BaseView
-
-LOCAL_FQDN = socket.getfqdn()
-VERSION = pkg_resources.get_distribution('ichnaea').version
-VERSION_FILE = os.path.join(ROOT, 'version.json')
-VERSION_INFO = {
-    'commit': 'HEAD',
-    'source': 'https://github.com/mozilla/ichnaea',
-    'tag': 'master',
-    'version': VERSION,
-}
-
-if os.path.isfile(VERSION_FILE):
-    with open(VERSION_FILE, 'r') as fd:
-        data = simplejson.load(fd)
-    VERSION_INFO['commit'] = data.get('commit', None)
-    VERSION_INFO['tag'] = data.get('tag', None)
 
 
 def _check_timed(ping_function):