+++ CHANGES.rst
@@ -24,6 +24,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- Configure release for raven/sentry client.
+
 - Set stations source column to gnss on insert/update.
 
 - Use query observations to detect and blocklist stations.